De-fang ConvertToPFiles.m
ConvertToPFiles.m overwrites PlannerConfig.xml asnd deletes a bunch
of files that where not needed/wanted for earlier p-code only releases
of Planner. Since Planner is now distributed as m-files the function is
not really needed anymore. Do to its side effects of deleting/replacing
files like PlannerConfig.xml a user wanting to run it will need to edit
the really_convert_to_pfiles to be true before running it.

Signed-off-by: Sebastian Moeller <moeller0@gmx.de>